refactor condition matchers to a class
